requirements.txt
streamlit==1.41.1
langchain==0.3.16
langchain_openai==0.3.2
serpapi==0.1.5
dotenv-python==1.0.1
openai==1.60.2
google-search-results==2.4.2